+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Added support to Logical switch group actions
   - Added support to Logical switch actions
   - Added support to Firmware bundle actions
+  - Added support to Firmware driver actions and integrated it with firmware bundles in a resource called Firmware
   - Added diff output for updating resources
 
 ### 0.1.1